Added occupancy bounding box functionality to speed up volume rendering even faster
Also added slider for occupancy map block
resolution

+++ src/shaders/includes/raycast_volume.glsl
@@ -1,5 +1,5 @@
 uniform float u_raycast_threshold;  // Threshold value to determine if the ray has hit an object
-uniform int u_raycast_refinements;  // Number of refinements for raycasting precision
+uniform float u_raycast_refinements;  // Number of refinements for raycasting precision
 uniform float u_raycast_dithering;
 uniform sampler2D u_noisemap_data;
